WAYK-2562: initial HTTP bridge implementation
Add a new route for http bridging validated using a JWT.
This is similar to the one in Wayk Bastion, but a JWT signed with the provisioner key is used for permission.
The JWT currently only provides the destination to forward to.
This is basically a kind of HTTP reverse proxy.

Related ticket: https://devolutions.atlassian.net/browse/WAYK-2562