feat(installer): improved error handling and rollback in Windows installer (#397)
PowerShell configuration commands are now executed as custom actions instead of WixSilentExec. Errors are tracked and, if the installer is running with UI, an appropriate error message is shown to the user. For first time installs, if the installation fails, files that may have been created by the configuration process are cleaned up.

PowerShell command output is redirected to a temporary file; in the case of an error we provide the user the path to that file. A general command execution error will display a string error value. 

Custom actions are refactored slightly for consistency and readability:

- Internal functions now only return `void`, `BOOL`, or `HRESULT` where possible. Errors are always handled as `HRESULT` and other results (e.g. Win32 error codes, `LSTATUS`, null references) are converted to `HRESULT` and handled with the different WiX macros (e.g. `ExitOnWin32Error`).
- Consolidate on `WixGetProperty` instead of `MsiGetProperty` and be careful to release the resulting strings (`ReleaseStr`)
- Consolidate on `nullptr` instead of `NULL`

Issue: DGW-76
Issue: DGW-78