fix(dgw): ensure the hostname matches TLS certificate
Warning logs are ignored at this point (logger not yet initialized),
so it doesn’t really help. Since specifying a hostname not matching the
TLS subject name is a configuration error, we now return an error upon
loading the configuration.Log warnings are ignored at this point, so it
doesn’t really help.