fix(dgw): better support for ngrok free plan
Our installer is allowing the 0.0.0.0/0 CIDR by default because
premium plans need the IP restrictions to be configured or just
all external traffic. However this doesn’t play well with the free
plan. This patch is using a dirty trick to detect the free plan
and ignores the IP restriction configuration when it is detected.

Issue: DGW-134