fix(dgw): make recording policy enforcing more robust
Previously the policy wasn’t properly enforced when the session was
started after the recording stream was opened.
In practice it’s very unlikely to happen since all clients we are
aware of are opening the recording stream after the session is started.
However, at the Devolutions Gateway level, opening the recording stream
first is also a totally valid flow that we don’t forbid.
This patch is accounting for this by updating the recording policy
accordingly on recording manager side once the session is opened.