fix(agent): make UpdaterConf::enabled required when deserializing
Instead of special casing the "Enabled" property of the updater module
config, we requires it when deserializing from JSON, and if the config
of the updater is completely removed, we assume that it’s not enabled.

We keep the updater enabled in the default config file.

Future modules will be enabled by default when generating a fresh
config file as well, but we won’t enable them by default implicitly
when installing a newer version of the executable.
It’s easier to predict how the agent will behaves for both us and the
user.