perf(jetsocat,dgw): new flow control
The threshold to send the WINDOW ADJUST message is made dynamic.

The following formula is used to determine the threshold:

  threshold = initial_window_size ÷ (num_channels + 3) × 3

That is, the more open channels, the higher the threshold.

E.g.: for initial_window_size = 1000, a message will be sent if
- num_channel = 1 and window_size <= 750
- num_channel = 2 and window_size <= 600
- num_channel = 3 and window_size <= 500
- num_channel = 10 and window_size <= 230
- num_channel = 50 and window_size <= 56
- num_channel = 100 and window_size <= 29

A channel using a lot of the throughput will have to wait for a WINDOW
ADJUST message at some point, allowing the other channels to use a
higher share of the the available throughput. This system also helps
reducing the number of WINDOW ADJUST messages sent over the wire
by eliminating all "small" WINDOW ADJUST messages which overall are
irrelevants.