chore(openapi): update generators to 7.9.0
The config was also updated to drop RestSharp as a dependency in the
.NET client.

It appears however that mostly everything in the `Configuration` class is
ignored, including `RemoteCertificateValidationCallback`.

Instead, it’s possible to provide a `HttpClient` or `HttpClientHandler`
instance directly, with relevant options such as
`ServerCertificateCustomValidationCallback` configured as appropriate.

Issue: DEVEX-114