fix(jetsocat): preemptively install crypto provider on start
This is typically not required because we only enable the ring
backend, which is then automatically installed by rustls when
`ClientConfig::builder` is called later down the road. However, it may
be the case that during development the aws-lc-rs feature of rustls is
enabled (especially when building the whole workspace at once), causing
rustls to crash the process requesting to install a default crypto
provider. To smoothen the developer experience, we attempt to install
the ring crypto provider just before executing the command action. Our
CI is already ensuring that we are not shipping both crypto backends
by mistake.

Changelog: ignore