refactor(pedm): audit and refactor privileges module
Audit and refactor the privileges module to remove code found to be
causing undefined behaviors.
The token module was also partially rewritten accordingly.

To help ensuring the new code is correct and stays that way, tests were
added, and the CI was updated to run them with the LLVM memory
sanitizers and miri (for all the code that is not interfacing with
Windows API directly)

Notable abstractions added:

- Safer wide strings (str module): Re-export string types from the
 widestring crate, and provide extension traits to integrate with
 the windows-strings crate.

- RawBuffer (raw_buffer module): A convenient RAII wrapper around the
 Rust std global allocator.

- Win32Dst (dst module): Generic wrapper for Win32-style dynamically
 sized types.

Issue: DGW-222
Changelog: ignore