build: add RPM for gateway and agent (#1179)
This commit adds RPM packages for Gateway and Agent to the release
assets.

`tlk.ps1 -TlkVerb package -Platform linux` now generates both a deb and
an rpm for the specified product.

The rpm is generated with fpm, a Linux packaging tool.
In CI, no cache is used for the fpm dependency since the cache would not
persist across release jobs.

The RPM includes all the assets of the corresponding Debian package,
including the changelog, copyright, maintainer scripts, and
webapp/libxmf.so for Gateway.

The changelog is copied directly from Debian. A separate PR will be made
to convert it to an RPM-friendly changelog format.

Tested with RHEL 9 (glibc 2.34).