fix(dgw): fix illegal invocation for send method on streaming player websocket proxy (#1183)
Fix the problem when Websocket send is called, we have an illegal
invocation exception. The Websocket send method needs to be tied to the
Websocket instance instead of the proxy, that is, the `this` keywords
need to be bind to the correct object. This problem was introduced in
the previous effort to intercept the `close` method for Websocket

Changelog: ignore

---------

Co-authored-by: Benoît Cortier <3809077+CBenoit@users.noreply.github.com>