fix(pedm): copy the shell extension before registration (#1219)
Upgrading or uninstalling the PEDM shell extension is still problematic:
in my test, the restart manager killed Windows Explorer and then failed
to restart it properly. This caused a broken state on my machine that
was difficult to recover.

Instead, we copy the shell extension DLL before registering it. On
uninstall, the dll is moved to a temporary location and then marked for
removal on reboot. This is the same approach taken by Notepad++ to avoid
this issue.