build: restart systemd service on package upgrade (#1232)
This uses the generated code from dh_installsystemd to restart the Gateway service when it is upgraded via a deb install.

It also has a clean up commit for the Agent scripts. The Agent systemd service registration is currently incomplete. It is
not done at the application-level or in packaging. This removes the systemd logic from the maintainer scripts.