Fetch integer as i64
SQLite stores integers in 64 bit.

If we try to get i32, it is just [cast internally](https://docs.rs/libsql/0.9.2/src/libsql/rows.rs.html#169-177) as i32.

The logic for fallible int conversion is already written. This commit
makes use of it and makes it explicit that we are actually retrieving an i64 from the database before converting it to our model type.