Remove unnecessary unwrap
This one was a bit funny.

`query_opt` returns `Option<Row>`. It was doing `opt_row.map(|r| r.get::<_, Option<D>(0))`, which would be `Some(Some(D))` or `None` if the row was not found. With `unwrap_or_default`, this was returning `Some(D)` or `None`, which is perfectly well.

However, the actual column is not nullable so `opt_row.map(|r|
r.get::<_, D>(0))` is simpler.