fix(pedm): CryptCATAdminCalcHashFromFileHandle2 does not return error with NULL buffer (#1323)
The `CryptCATAdminCalcHashFromFileHandle2`
[documentation](https://learn.microsoft.com/en-us/windows/win32/api/mscat/nf-mscat-cryptcatadmincalchashfromfilehandle2)
says:


>Return value
>
>[...]
>
>ERROR_INSUFFICIENT_BUFFER | The buffer pointed to by
the pbHash parameter was not NULL but was not large enough to be
written. The correct size of the required buffer is contained in the
value pointed to by the pcbHash parameter.

Ergo, the function will **not** return an error when passing a `NULL`
value for `pbHash`, it will succeed. This is confusing because other
similar Win32 APIs do not work the same way.