feat(pedm): explicitly support elevating .msi files (#1325)
By inspecting elevated .msi files launched from Explorer, we see that
Explorer invokes %systemroot%\system32\msiexec, with the command line
`"%systemroot%\system32\msiexec" /i "{path-to-msi}"`.

We achieve the same in the PEDM module by using the same command if the
file extension is .msi. The .msi extension is already being trapped by
the shell extension, but previously we would call `CreateProcess` on the
.msi causing "file is not a valid Win32 executable".

Additionally, I polished and added @awakecoding's original
"ExplorerCommand.ps1" script. This is useful for debugging, where the
installed shell extension can't be overwritten as it's in use by
Explorer. Install Devolutions Agent with the PEDM feature; then use the
PowerShell to unregistered the shell extension. Swap in your development
or debug DLL, then re-register the shell extension. If you supply the
.pdb file as well it's possible to attach the debugger to Explorer.exe
and debug the shell extension code.