feat(pedm): add version check API and signature check API (#1343)
Add a new policy option `target_must_be_signed`; this forces the target
executable to have a valid auhenticode signature as per the policy.

Realistically, this should be a "rule" so I'm probably introducing some
slight technical debt here but I believe it's a worthwhile feature and a
low-hanging fruit we can squeeze in for the release.

Add the application (Agent) version to the "About" API endpoint. This is
for future proofing: RDM needs a way to know what version of PEDM he's
talking to, in case we add/remove/change APIs.

The implementation is in Win32 and I believe it to be correct, but it
could use some close scrutiny.

I generated the OpenAPI in a. separate commit for easier reviewing.