fix(pedm): ensure get_last_request_time has a value (#1352)
The comments for
[AboutData](https://github.com/Devolutions/devolutions-gateway/blob/668e0b43edd5ee2ce4cbf4692920c09d732eeb85/crates/devolutions-pedm/src/model.rs#L7)
say

> This can be `None` if `/about` is the first request made.

Which is true and correct; but is causing a crash in the .NET OpenAPI
client if /about is the first request made.

For whatever reason, the .NET OpenAPI
[generator](https://github.com/Devolutions/devolutions-gateway/blob/668e0b43edd5ee2ce4cbf4692920c09d732eeb85/crates/devolutions-pedm/openapi/dotnet-client/src/Devolutions.Pedm.Client/Model/AboutData.cs#L74)
is not creating a nullable property.

This may be related to the issues I had in #1164, but I don't have time
to dig into it now. Since this is non-critical / non-user facing; I just
force a value here for the first request (the UNIX epoch seems
reasonable).