fix(dgw): set default value of TlsVerifyStrict to false (#1419)
Previously, strict TLS verification was performed even when the
TlsVerifyStrict key was absent from the configuration file.

From now on, if this key is missing, it will default to
"TlsVerifyStrict": false.

This change ensures that existing users who are currently using improper
certificates will not be affected. At the same time, newly generated
configuration files will continue to include "TlsVerifyStrict": true by
default, encouraging using proper certificates from the start.

New users can still opt out of strict verification by explicitly setting
the value to false or removing the key entirely if they are willing to
accept potential compatibility issues with some clients, such as Chrome
or macOS.

A warning will be logged if the option is disabled as it may hide latent
issues.
Hopefully, this lead the user to enable the option, and fix the
underlying certificate issue if necessary.