fix(webapp)!: add manual clipboard mode for RDP, VNC and ARD (#1459)
The original auto clipboard mode, can now be enabled/disabled in the connection form.
It's available only for browsers based on _Blink_ engine (e.g.: Chrome).

For others, auto clipboard mode is always disabled.

When disabled, two new buttons are showed in the toolbar: _Save Clipboard_ and _Send Clipboard_.
These buttons allow the user to manually send the content of the client host clipboard, or receive from
the server clipboard.

Note that Firefox used to have a partially working autoclipboard mode, but it was
completely removed because the cases where it failed were confusing more than helpful.
The auto clipboard mode is the most convenient, but the manual clipboard mode
has a predictable behavior on all platforms.