ci: fix PowerShell .tar validation (#1505)
We saw that a .tar archive was corrupted with the warning "There are
data after the archive [sic]". This seems like something transient, but
validation didn't catch it.

`tar -t {path}` will not work as expected (at least on Windows); it's
needed to pass `f` to specify the file path. Otherwise tar falls back on
`\\.\tape0)`(!).

Regardless, when validating tar does not set a non-zero error code for
this kind of issue.

Instead, use `7z -t` which properly sets the error code and allows us to
bail out.