fix(jetsocat): resolve channel open -> start tight race condition
Fixes a race condition in the JMUX proxy where data could arrive  from a
peer before the listener had a chance to send the channel start request,
causing data to be lost.

When a channel is opened, there's a small window where incoming
data from the peer arrives before the corresponding data sender is
established. Previously, this data was dropped with a warning.

Issue: ARC-323